Created models module. Removed all the large blocks commented out.
