+++ demos/models/__init__.py
@@ -1,0 +1,10 @@
+from .aging import *
+from .laborforce import *
+from .household_reorg import *
+from .marriage import *
+from .kids_moving import *
+from .mortality import *
+from .birth import *
+from .education import *
+
+from .main import * # TODO: Refactor this
